pyproject.toml

[project]
name = "upgrade-advisor"
version = "0.1.0"
description = "An AI-powered GitHub App that suggests dependency upgrades for your repositories."
authors = [
    { name = "thatgeeman", email = "thatgeeman@users.noreply.github.com" },
]
readme = "README.md"
requires-python = ">=3.10,<3.15"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "smolagents[mcp] (>=1.22.0,<2.0.0)",
    "mcp (>=1.21.1,<2.0.0)",
    "pytest (>=9.0.1,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "anyio (>=4.11.0,<5.0.0)",
    "transformers (>=4.57.1,<5.0.0)",
    "typer[all] (>=0.20.0,<0.21.0)",
    "gradio (>=5.49.1,<6.0.0)",
    "ddgs (>=9.9.1,<10.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = true
packages = [{ include = "src/upgrade_advisor" }]

# extend to ruff.toml
[tool.ruff]
extend = "ruff.toml"
